Rename ::types to ::common
